rename ingredients.slices to ingredients.slice
