+++ src/components/ingredient-details/ingredient-details.tsx
@@ -3,7 +3,7 @@ import { Preloader } from '../ui/preloader';
 import { IngredientDetailsUI } from '../ui/ingredient-details';
 import { useParams } from 'react-router-dom';
 import { useSelector } from '../../services/store';
-import { ingredientsSelector } from '../../slices/ingredients.slices';
+import { ingredientsSelector } from '../../slices/ingredients.slice';
 
 export const IngredientDetails: FC = () => {
   const { id } = useParams();
