+++ src/pages/constructor-page/constructor-page.tsx
@@ -6,7 +6,7 @@ import { BurgerIngredients } from '../../components';
 import { BurgerConstructor } from '../../components';
 import { Preloader } from '../../components/ui';
 import { FC } from 'react';
-import { isIngredientsLoadedSelector } from '../../slices/ingredients.slices';
+import { isIngredientsLoadedSelector } from '../../slices/ingredients.slice';
 
 export const ConstructorPage: FC = () => {
   const isIngredientsLoading = useSelector(isIngredientsLoadedSelector);
